mant: corrigindo diretorios e arrumando graficos
