+++ tupan/src/app/visualizar-alertas/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { MenuLateral } from "@/components/menu-lateral";
+import { MenuLateral } from "@/components/menu/lateral";
 import { NavTop } from "@/components/nav-top";
 import React, { useEffect, useState } from "react";
 import { Alerta } from "@/components/alertas";
