+++ tupan/src/app/visualizar-usuario/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { MenuLateral } from "@/components/menu-lateral";
+import { MenuLateral } from "@/components/menu/lateral";
 import { NavTop } from "@/components/nav-top";
 
 import React from 'react';
