+++ tupan/src/components/graficos/export.tsx
@@ -1,0 +1,5 @@
+import BarChart from "./barra/barchart";
+import ColumnChart from "./coluna/clomun";
+import LineChart from "./linha";
+
+export { BarChart, ColumnChart, LineChart };
