+++ tupan/src/components/tabela/tabela-parametros.tsx
