+++ tupan/src/hooks/parametros/adicionarParametro.ts
