+++ tupan/src/hooks/parametros/export.ts
@@ -1,0 +1,6 @@
+import { useCreateParametro } from "./adicionarParametro";
+import { useDeleteParametro } from "./deletarParametro";
+import { useUpdateParametro } from "./atualizarParametro";
+import { useGetParametros } from "./receberParametro";
+
+export { useCreateParametro, useDeleteParametro, useGetParametros, useUpdateParametro}
