+++ tupan/src/hooks/parametros/receberParametro.ts
